Implement onboarding client component and refactor onboarding page for user profile checks
